requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
json5=0.9.6
music21=7.1.0
numpy=1.21.4
python-slugify=5.0.2
pydantic=1.8.2